% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_utils.R
\name{assign_iucn}
\alias{assign_iucn}
\title{assign_iucn()}
\usage{
assign_iucn(change, criteria = c("A2", "A1")[1])
}
\arguments{
\item{change}{measured change over 3 generations}

\item{criteria}{A1 or A2 for decline}
}
\value{
IUCN status
}
\description{
Function to assign IUCN status based on decline and criteria
}
