% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_utils.R
\name{dfidx}
\alias{dfidx}
\title{dfidx()}
\usage{
dfidx(jara, run = "obs", backcast.na = NULL, scale.q = FALSE)
}
\arguments{
\item{jara}{fit_jara output}

\item{run}{name qualifier of data.frame}

\item{backcast.na}{option to remove backcast prediction prior to a year}
}
\value{
data.frame
}
\description{
Function to compile fits into obs and fit data.frames
}
\author{
Henning Winker (JRC-EC)
}
