% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_jara.R
\name{fit_jara}
\alias{fit_jara}
\title{fit_jara()}
\usage{
fit_jara(
  jarainput,
  ni = 11000,
  nt = 2,
  nb = 1000,
  nc = 2,
  save.jara = FALSE,
  save.all = FALSE,
  save.csvs = FALSE,
  save.jarafile = TRUE,
  output.dir = getwd(),
  quickmcmc = FALSE
)
}
\arguments{
\item{jarainput}{List of input variables as output by build_jara()
MCMC settings}

\item{ni}{number of iterations}

\item{nt}{thinning interval of saved iterations}

\item{nb}{burn-in}

\item{nc}{number of mcmc chains}

\item{save.jara}{= FALSE}

\item{save.all}{= FALSE}

\item{output.dir}{path to save plot. default is getwd()}
}
\value{
A result list containing estimates of JARA model input, settings and results
}
\description{
Fits JARA model in JAGS and produce output object as list()
}
