% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_utils.R
\name{iucn}
\alias{iucn}
\title{iucn}
\usage{
iucn(jara, criteria = c("A2", "A1")[1])
}
\arguments{
\item{jara}{output list from fit_jara}

\item{criteria}{option to choose between IUCN A1 or A2 thresholds (A2 is default)}
}
\value{
IUCN classification
}
\description{
IUCN computes IUCN status for A1 or A2
}
\author{
Henning Winker, Richard Sherley and Nathan Pacoureau
}
