% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_utils.R
\name{iucn_estimators}
\alias{iucn_estimators}
\title{iucn_estimators()}
\usage{
iucn_estimators(I = NULL, GL = 10, criteria = c("A2", "A1")[1])
}
\arguments{
\item{I}{obserseved abundance index data.frame(yr,y)}

\item{GL}{generation length}

\item{criteria}{A1 or A2 for decline}
}
\value{
change over 3 x GL for methods p2 and reg
}
\description{
Current estimators based on two points (p2) and log-linear regression (reg)
}
