% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_utils.R
\name{iucn_frame}
\alias{iucn_frame}
\title{iucn_frame}
\usage{
iucn_frame(
  xylim = c(100, 1),
  plot.cex = 1,
  legend.cex = 0.9,
  criteria = c("A2", "A1")[1],
  iucn.cols = TRUE,
  add = FALSE
)
}
\arguments{
\item{plot.cex}{cex graphic option}

\item{legend.cex}{lengend size cex graphic option}

\item{criteria}{option to choose between IUCN A1 or A2 thresholds (A2 is default)}

\item{iucn.cols}{to use iucn color recommendation or a brighter version if FALSE}

\item{add}{if TRUE par is not called to enable manual multiplots}

\item{xymax}{determines upper x and y lims}

\item{width}{plot width}

\item{height}{plot hight}
}
\description{
Empty iucn color plot
}
\author{
Henning Winker, Richard Sherley and Nathan Pacoureau
}
