% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara2jags.R
\name{jara2jags}
\alias{jara2jags}
\title{Function to write JARA model in JAGS}
\usage{
jara2jags(jarainput)
}
\arguments{
\item{jarainput}{JARA input object from build_jara()}
}
\description{
Writes JAGS code for JABBA into a temporary directory
}
