% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_hindcast.R
\name{jara_hindcast}
\alias{jara_hindcast}
\title{JARA hindcasting function}
\usage{
jara_hindcast(
  jarainput,
  ni = 9000,
  nt = 2,
  nb = 2000,
  nc = 3,
  peels = 0:5,
  save.jara = TRUE,
  save.all = FALSE,
  save.csvs = FALSE,
  output.dir = getwd(),
  save.jarafile = TRUE,
  save.hc = TRUE,
  plotall = FALSE,
  speedup = TRUE,
  rm.yr = TRUE
)
}
\arguments{
\item{jarainput}{List of input variables as output by build_jabba()
MCMC settings}

\item{ni}{number of iterations}

\item{nt}{thinning interval of saved iterations}

\item{nb}{burn-in}

\item{nc}{number of mcmc chains}

\item{peels}{sequence of retro spective peels default 0:5}

\item{save.jara}{saves jara list as .rdata to output.dir}

\item{save.all}{saves the all posteriors as .rdata to output.dir (big file)}

\item{save.csvs}{writes results into csv to output.dir}

\item{output.dir}{path to save plot. default is getwd()}

\item{save.jarafile}{saves jara model and convergence stats as .txt file (default = TRUE)}

\item{save.hc}{Save hindcast list output as .rdata}

\item{plotall}{if TRUE makes jara_plots() for each run}

\item{speedup}{Reduces MCMC after setting runs 2+ inits to first "full" reference run}

\item{rm.yr}{if TRUE remove last year, if FALSE replace obs with NA}

\item{saves}{the all posteriors as .rdata to output.dir (big file)}
}
\value{
A result list containing estimates of JARA model input, settings and results

hc containing estimates of key joint results from all hindcast run
}
\description{
Wrapper to coduct histcasts for retrospective analysis and cross-validation
}
