% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_plots.R
\name{jara_plots}
\alias{jara_plots}
\title{wrapper jara_plots function}
\usage{
jara_plots(jara, output.dir = getwd(), as.png = TRUE, statusplot = "kobe")
}
\arguments{
\item{jara}{output list from fit_jara}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}
}
\description{
plots all routine JARA plots to output.dir if as.png=TRUE (default)
}
