% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointindex.R
\name{jointindex}
\alias{jointindex}
\title{jointindex()}
\usage{
jointindex(jara, credibility = 0.95)
}
\arguments{
\item{jara}{output list from fit_jara}

\item{credibility}{Confidence interval credibility interval with default of 0.95}
}
\value{
data.frame
}
\description{
Generates a joint index from the population and joint residuals
}
\author{
Henning Winker
}
