% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_plots.R
\name{jrplot_changes}
\alias{jrplot_changes}
\title{jrplot_changes}
\usage{
jrplot_changes(
  jara,
  output.dir = getwd(),
  as.png = FALSE,
  width = 5,
  height = 4.5,
  ylab = "Density",
  xlab = "Annual rate of change (\%)",
  plot.cex = 1,
  legend.cex = 0.9,
  add = FALSE
)
}
\arguments{
\item{jara}{output list from fit_jara}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{width}{plot width}

\item{height}{plot hight}

\item{ylab}{option to change y-axis label}

\item{xlab}{option to change x-axis label}

\item{plot.cex}{cex graphic option}

\item{legend.cex}{legend sizr graphic option}

\item{add}{if TRUE par is not called to enable manual multiplots}
}
\description{
Plot mean rates of change (%) over 1, 2 anf 3 Generation lengths
}
\author{
Henning Winker and Richard Sherley
}
