% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_plots.R
\name{jrplot_fits}
\alias{jrplot_fits}
\title{Plot of  fitted abundance indices}
\usage{
jrplot_fits(
  jara,
  output.dir = getwd(),
  as.png = FALSE,
  single.plots = FALSE,
  width = NULL,
  height = NULL,
  indices = "all"
)
}
\arguments{
\item{jara}{output list from fit_jara}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{single.plots}{if TRUE plot invidual fits else make multiplot}

\item{width}{plot width}

\item{height}{plot hight}
}
\description{
Plots observed and fitted indices with expexted CIs (dark grey)
}
