% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_plots.R
\name{jrplot_indices}
\alias{jrplot_indices}
\title{jrplot_indices}
\usage{
jrplot_indices(
  jarainput,
  output.dir = getwd(),
  as.png = FALSE,
  width = 5,
  height = 4.5,
  plot.cex = 1,
  xlab = "Year",
  ylab = "Abundance index",
  add = FALSE,
  add.legend = TRUE,
  cols = NULL
)
}
\arguments{
\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{width}{plot width}

\item{height}{plot hight}

\item{plot.cex}{cex graphic option}

\item{xlab}{option to change x-axis label}

\item{ylab}{option to change y-axis label}

\item{add}{if TRUE par is not called to enable manual multiplots}

\item{add.legend}{option to add legend}

\item{cols}{option to choose own colour palette}

\item{jara}{output list from fit_jara}

\item{criteria}{A1 or A2 for decline}
}
\description{
Plot mean rates of change (%) over 1, 2 anf 3 Generation lengths
}
