% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_plots.R
\name{jrplot_iucn}
\alias{jrplot_iucn}
\title{jrplot_iucn}
\usage{
jrplot_iucn(
  jara,
  output.dir = getwd(),
  as.png = FALSE,
  width = 5,
  height = 4.5,
  plot.cex = 1,
  criteria = c("A2", "A1")[1],
  iucn.cols = TRUE,
  add = FALSE,
  Plot = TRUE
)
}
\arguments{
\item{jara}{output list from fit_jara}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{width}{plot width}

\item{height}{plot hight}

\item{criteria}{A1 or A2 for decline}

\item{add}{if TRUE par is not called to enable manual multiplots}
}
\value{
IUCN classification
}
\description{
IUCN posterior plot for A1 or A2
}
