% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_plots.R
\name{jrplot_jointindex}
\alias{jrplot_jointindex}
\title{jrplot_jointindex}
\usage{
jrplot_jointindex(
  jara,
  credibility = 0.95,
  output.dir = getwd(),
  as.png = FALSE,
  width = 5,
  height = 4.5,
  ylab = "Joint-Index",
  show.trend = TRUE,
  xlab = "Year",
  xlim = NULL,
  plot.cex = 0.8,
  add = FALSE
)
}
\arguments{
\item{jara}{output jara}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{width}{plot width}

\item{height}{plot hight}

\item{ylab}{option to change y-axis label}

\item{xlab}{option to change x-axis label}

\item{xlim}{plotting optio}

\item{plot.cex}{cex graphic option}

\item{add}{if TRUE par is not called to enable manual multiplots}

\item{legend.pos}{Choose legend position}

\item{legend.cex}{legend sizr graphic option}
}
\description{
Plots timeblock effect or change
}
\author{
Henning Winker
}
