% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_plots.R
\name{jrplot_logfits}
\alias{jrplot_logfits}
\title{log(index) fits}
\usage{
jrplot_logfits(
  jara,
  output.dir = getwd(),
  as.png = FALSE,
  single.plots = FALSE,
  width = NULL,
  height = NULL,
  ylab = "Log(index)",
  xlab = "Year",
  indices = "all",
  add = FALSE
)
}
\arguments{
\item{jara}{output list from fit_jara}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{single.plots}{if TRUE plot invidual fits else make multiplot}

\item{width}{plot width}

\item{height}{plot hight}

\item{ylab}{option to change y-axis label}

\item{xlab}{option to change x-axis label}

\item{indices}{names of indices to plot (default = "all")}

\item{add}{if TRUE par is not called to enable manual multiplots}
}
\description{
Plot of fitted CPUE indices on log-scale (r4ss-style)
}
