% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_plots.R
\name{jrplot_poptrj}
\alias{jrplot_poptrj}
\title{jrplot_poptrj()}
\usage{
jrplot_poptrj(
  jara,
  output.dir = getwd(),
  as.png = FALSE,
  width = 5,
  height = 4.5,
  plot.cex = 1,
  add = FALSE
)
}
\arguments{
\item{jara}{output list from fit_jara}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{width}{plot width}

\item{height}{plot hight}

\item{add}{if TRUE par is not called to enable manual multiplots}
}
\description{
Plots the estimated population trajector relative to GL
}
