% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_plots.R
\name{jrplot_r}
\alias{jrplot_r}
\title{jrplot_r}
\usage{
jrplot_r(
  jara,
  output.dir = getwd(),
  as.png = FALSE,
  width = 5,
  height = 4.5,
  xlab = "r",
  ylab = "Density",
  plot.cex = 1,
  add = FALSE
)
}
\arguments{
\item{jara}{output list from fit_jara}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{width}{plot width}

\item{height}{plot hight}

\item{xlab}{option to change x-axis label}

\item{ylab}{option to change y-axis label}

\item{plot.cex}{cex graphic option}

\item{add}{if TRUE par is not called to enable manual multiplots}
}
\description{
Plot mean rates r = log(lamda) over 1, 2 anf 3 Generation lengths
}
