% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_plots.R
\name{jrplot_residuals}
\alias{jrplot_residuals}
\title{joint residual plot}
\usage{
jrplot_residuals(
  jara,
  output.dir = getwd(),
  as.png = FALSE,
  add = FALSE,
  ylab = "log Residuals",
  xlab = "Year",
  width = 5,
  height = 3.5,
  cols = NULL
)
}
\arguments{
\item{jara}{output list from fit_jara}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{add}{if true don't call par() to allow construction of multiplots}

\item{ylab}{option to change y-axis label}

\item{xlab}{option to change x-axis label}

\item{width}{plot width}

\item{height}{plot hight}

\item{cols}{option to add colour palette}
}
\description{
plots residuals for all indices as boxplot with a loess showing systematic trends
}
