% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_plots.R
\name{jrplot_retrobias}
\alias{jrplot_retrobias}
\title{jrplot_retrobias() to plot retrospective pattern}
\usage{
jrplot_retrobias(
  hc,
  output.dir = getwd(),
  as.png = FALSE,
  width = NULL,
  height = NULL,
  Xlim = NULL,
  cols = NULL,
  legend.loc = "topright",
  show.rho = TRUE
)
}
\arguments{
\item{hc}{output from jabba_hindast()}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{width}{plot width}

\item{height}{plot hight}

\item{Xlim}{allows to "zoom-in" requires speficiation Xlim=c(first.yr,last.yr)}

\item{cols}{option to add colour palette}

\item{legend.loc}{location of legend}

\item{show.rho}{shows rho statistic in plot}
}
\value{
Mohn's rho statistic for several quantaties
}
\description{
Plots retrospective pattern of B, F, BBmsy, FFmsy, BB0 and SP #'
}
