% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_plots.R
\name{jrplot_retroiucn}
\alias{jrplot_retroiucn}
\title{jrplot_retroiucn}
\usage{
jrplot_retroiucn(
  hc,
  output.dir = getwd(),
  as.png = FALSE,
  width = 5,
  height = 4.5,
  plot.cex = 1,
  criteria = c("A2", "A1")[1],
  iucn.cols = TRUE,
  ylab = "Change (\%)",
  xlab = "Year",
  add = FALSE,
  Plot = TRUE,
  add.legend = TRUE
)
}
\arguments{
\item{hc}{output list from jara_hindcast()}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{width}{plot width}

\item{height}{plot hight}

\item{plot.cex}{cex graphic option}

\item{criteria}{option to choose between IUCN A1 or A2 thresholds (A2 is default)}

\item{iucn.cols}{to use iucn color recommendation or a brighter version if FALSE}

\item{ylab}{option to change y-axis label}

\item{xlab}{option to change x-axis label}

\item{add}{if TRUE par is not called to enable manual multiplots}

\item{Plot}{if FALSE then only threat status stats are returned}

\item{add.legend}{option add legend}
}
\value{
Retrospectice IUCN classification
}
\description{
Retrospective IUCN posterior plot for A1 or A2
}
\author{
Henning Winker, Richard Sherley and Nathan Pacoureau
}
