% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_plots.R
\name{jrplot_state}
\alias{jrplot_state}
\title{jrplot_state}
\usage{
jrplot_state(
  jara,
  type = NULL,
  ref.yr = NULL,
  extinction = 0.01,
  credibility = 0.95,
  output.dir = getwd(),
  as.png = FALSE,
  width = 5,
  height = 4.5,
  ylab = "Density",
  xlab = "Relative state",
  plot.cex = 1,
  legend.cex = 0.9,
  add = FALSE
)
}
\arguments{
\item{jara}{output list from fit_jara}

\item{type}{final year type = c("current","projected","both")}

\item{ref.yr}{year(s) used as reference; default is avg. first 3 years}

\item{extinction}{threshold for extiction classification default or ref.yr}

\item{credibility}{Confidence interval credibility calue with default of 0.95}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{width}{plot width}

\item{height}{plot hight}

\item{ylab}{option to change y-axis label}

\item{xlab}{option to change x-axis label}

\item{plot.cex}{cex graphic option}

\item{legend.cex}{legend size graphic option}

\item{add}{if TRUE par is not called to enable manual multiplots}
}
\value{
Relative state
}
\description{
Plots current or projected population change relative to the first year
}
\author{
Henning Winker
}
