% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_plots.R
\name{jrplot_timeblock}
\alias{jrplot_timeblock}
\title{jrplot_timeblock}
\usage{
jrplot_timeblock(
  jara,
  type = c("change", "r"),
  credibility = 0.95,
  probability = TRUE,
  output = TRUE,
  output.dir = getwd(),
  as.png = FALSE,
  width = 5,
  height = 4.5,
  ylab = "Density",
  xlab = NULL,
  xlim = NULL,
  plot.cex = 1,
  legend.cex = 0.9,
  legend.pos = "topright",
  add = FALSE
)
}
\arguments{
\item{jara}{output list from fit_jara}

\item{type}{"change"}

\item{credibility}{Confidence interval credibility calue with default of 0.95}

\item{probability}{show probility P(Effect>0) in plot}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{width}{plot width}

\item{height}{plot hight}

\item{ylab}{option to change y-axis label}

\item{xlab}{option to change x-axis label}

\item{xlim}{plotting optio}

\item{plot.cex}{cex graphic option}

\item{legend.cex}{legend sizr graphic option}

\item{legend.pos}{Choose legend position}

\item{add}{if TRUE par is not called to enable manual multiplots}
}
\description{
Plots timeblock effect or change
}
\author{
Henning Winker
}
