% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_plots.R
\name{jrplot_trjfit}
\alias{jrplot_trjfit}
\title{jrplot_trjfit()}
\usage{
jrplot_trjfit(
  jara,
  output.dir = getwd(),
  as.png = FALSE,
  width = 5,
  height = 4.5,
  plot.cex = 1,
  add = FALSE,
  indices = "all"
)
}
\arguments{
\item{jara}{output list from fit_jara}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{width}{plot width}

\item{height}{plot hight}

\item{plot.cex}{cex graphic option}

\item{add}{if TRUE par is not called to enable manual multiplots}

\item{indices}{names of indices to plot (default = "all")}
}
\description{
Plots the estimated and predicted population trajectors on the same scale
}
