% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_utils.R
\name{rqpois}
\alias{rqpois}
\title{rqpois()}
\usage{
rqpois(n, lambda, phi)
}
\arguments{
\item{n}{number of random observations}

\item{lambda}{vector of (non-negative) expected means}

\item{phi}{over-dispersion (variance) parameter}
}
\value{
random quasi-poisson observations
}
\description{
Random generator for quasi-poisson variables
}
