% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_jara.R
\name{build_jara}
\alias{build_jara}
\title{build_jara function}
\usage{
build_jara(
  I = NULL,
  se = NULL,
  assessment = "species.X",
  scenario = "s1",
  model.type = c("relative", "census")[1],
  GL = NULL,
  start.year = NA,
  end.year = NA,
  sigma.obs.est = TRUE,
  fixed.obsE = NULL,
  sigma.proc.fixed = FALSE,
  proc.pen = NULL,
  Klim = FALSE,
  K.manual = FALSE,
  prjr.type = c("all", "GL1", "year")[1],
  proj.stoch = FALSE
)
}
\arguments{
\item{I}{abundance indices/counts,  require data.frame(year, I.1,I.2,...,I.N)}

\item{se}{optional log standard error (CV) time series,requires data.frame(year, se.1,se.2,...,se.N)}

\item{assessment}{= "species.X",}

\item{scenario}{= "s1",}

\item{model.type}{abundance data type c("relative","census")}

\item{GL}{Generation length (default n.years/3)}

\item{start.year}{subsetting option for start year}

\item{end.year}{subsetting option for end year}

\item{sigma.obs.est}{Estimates additional variance}

\item{fixed.obsE}{minimum plausible process error (fixed)}

\item{sigma.proc.fixed}{option to fix the process error (default FALSE)}

\item{Klim}{penalty to restrict extrem values during projections c(TRUE, FALSE)}

\item{K.manual}{option to specify a carrying capacity for each census time series}

\item{prjr.type}{# rate of change for c("all","GL1", years), "all" is default}

\item{proj.stoch}{allows for projections with process error c(TRUE, FALSE), FALSE is default}
}
\value{
List to be used as data input to JARA JAGS model
}
\description{
Creates a data list with JABBA input and settings to be passed to fit_jara()
}
