% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_jara.R
\name{fit_jara}
\alias{fit_jara}
\title{fit_jara()}
\usage{
fit_jara(
  jarainput,
  credibility = 0.95,
  ni = 9000,
  nt = 2,
  nb = 2000,
  nc = 3,
  save.jara = FALSE,
  save.all = FALSE,
  save.csvs = FALSE,
  save.jarafile = TRUE,
  peels = NULL,
  output.dir = getwd(),
  quickmcmc = FALSE
)
}
\arguments{
\item{jarainput}{List of input variables as output by build_jara()}

\item{credibility}{Confidence Credibility Value with default of 0.95 
MCMC settings}

\item{ni}{number of iterations}

\item{nt}{thinning interval of saved iterations}

\item{nb}{burn-in}

\item{nc}{number of mcmc chains}

\item{save.jara}{saves jara list as .rdata to output.dir}

\item{save.all}{saves the all posteriors as .rdata to output.dir (big file)}

\item{save.csvs}{writes results into csv to output.dir}

\item{save.jarafile}{saves jara model and convergence stats as .txt file (default = TRUE)}

\item{peels}{= NULL, # retrospective 'peel' option}

\item{output.dir}{path to save plot. default is getwd()}

\item{quickmcmc}{option "test run" jara with short (fast) mcmc chains}

\item{saves}{the all posteriors as .rdata to output.dir (big file)}
}
\value{
A result list containing estimates of JARA model input, settings and results
}
\description{
Fits JARA model in JAGS and produce output object as list()
}
\examples{
data(jaradata)
inp = jrdat$Yellowspot_skate
buildjr <- build_jara(I = inp$I,se=inp$SE,GL=12)
fitjr <- fit_jara(buildjr)
jrplot_fits(fitjr)
jrplot_trjfit(fitjr)
jrplot_poptrj(fitjr)
jrplot_changes(fitjr)
jrplot_state(fitjr)
jrplot_iucn(fitjr)
}
\author{
Henning Winker and Richard Sherley
}
