% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara2jags.R
\name{jara2jags}
\alias{jara2jags}
\title{jara2jags()}
\usage{
jara2jags(jarainput)
}
\arguments{
\item{jarainput}{JARA input object from build_jara()}
}
\description{
Writes JAGS code for JARA into a temporary directory
}
\author{
Henning Winker
}
