% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrdata_simulator.R
\name{jrdat_simulator}
\alias{jrdat_simulator}
\title{jrdata_simulator()}
\usage{
jrdat_simulator(
  n0 = 1000,
  yrs = 40,
  r.mu = runif(1, -0.05, 0),
  proc.pop = 0.1,
  AR1.pop = 0.3,
  proc.imp = 0.05,
  AR1.imp = 0.8,
  CV.obs = 0.15,
  GL = 10,
  criteria = c("A2", "A1")[1],
  Plot = TRUE,
  add = FALSE
)
}
\arguments{
\item{n0}{start population size}

\item{yrs}{number of simulated years with observations}

\item{r.mu}{mean rate of change (default random uniform number)}

\item{proc.pop}{lognormal process error for natural population variation}

\item{AR1.pop}{1st order autoregressive error on poplation size}

\item{proc.imp}{lognormal process error on (anthropogenic) impact}

\item{AR1.imp}{1st order autoregressive error on impact}

\item{CV.obs}{observation error (~ log.sd)}

\item{GL}{Generation length to measure decline over 3 x GL}

\item{criteria}{A1 or A2 for decline}

\item{Plot}{Shows plot with with simulated data and processes}

\item{add}{if TRUE par is not called to enable manual multiplots}
}
\value{
simulated data and 'true' decline metrics
}
\description{
simulates observed and "true" abundance time series
}
\examples{
simdat <- jrdat_simulator(yrs=40,GL=15)
inpsim <- build_jara(I = simdat$I,GL=simdat$GL)
fitsim <- fit_jara(inpsim)
jrpar(mfrow=c(1,2))
jrplot_poptrj(fitsim,add=TRUE)
lines(simdat$N,col="cyan3",lwd=2)
jrplot_iucn(fitsim,add=TRUE)
abline(v=simdat$perc.change,col="cyan3",lwd=2)
}
\author{
Henning Winker (JRC-EC, Ispra)
}
