% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_plots.R
\name{jrpar}
\alias{jrpar}
\title{jrplot_pars()}
\usage{
jrpar(
  mfrow = c(1, 1),
  plot.cex = 1,
  mai = NULL,
  omi = c(0, 0, 0, 0),
  labs = TRUE
)
}
\arguments{
\item{mfrow}{determines plot frame set up}

\item{plot.cex}{cex graphic option}

\item{mai}{graphical par for plot margins}

\item{omi}{graphical par for outer plot margins}
}
\description{
Set the par() to options suitable for JARA multi plots
}
