% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_plots.R
\name{jrplot_iucn}
\alias{jrplot_iucn}
\title{jrplot_iucn}
\usage{
jrplot_iucn(
  jara,
  output.dir = getwd(),
  as.png = FALSE,
  width = 5,
  height = 4.5,
  plot.cex = 1,
  xlim = NULL,
  ylimadj = 1.1,
  legend.cex = 0.9,
  criteria = c("A2", "A1")[1],
  iucn.cols = TRUE,
  ylab = "Density",
  xlab = "Change (\%)",
  add = FALSE,
  Plot = TRUE
)
}
\arguments{
\item{jara}{output list from fit_jara}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{width}{plot width}

\item{height}{plot hight}

\item{plot.cex}{cex graphic option}

\item{xlim}{graphic option c(min,max)}

\item{ylimadj}{multiplier to adjust upper ylim}

\item{legend.cex}{lengend size cex graphic option}

\item{criteria}{option to choose between IUCN A1 or A2 thresholds (A2 is default)}

\item{iucn.cols}{to use iucn color recommendation or a brighter version if FALSE}

\item{ylab}{option to change y-axis label}

\item{xlab}{option to change x-axis label}

\item{add}{if TRUE par is not called to enable manual multiplots}

\item{Plot}{if FALSE then only threat status stats are returned}
}
\value{
IUCN classification
}
\description{
IUCN posterior plot for A1 or A2
}
\author{
Henning Winker, Richard Sherley and Nathan Pacoureau
}
