% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_plots.R
\name{jrplot_poptrj}
\alias{jrplot_poptrj}
\title{jrplot_poptrj()}
\usage{
jrplot_poptrj(
  jara,
  plotGL = NULL,
  output.dir = getwd(),
  as.png = FALSE,
  width = 5,
  height = 4.5,
  xlab = "Year",
  ylab = "Abundance",
  plot.cex = 1,
  add = FALSE
)
}
\arguments{
\item{jara}{output list from fit_jara}

\item{plotGL}{TRUE/FALSE indicates Generation Length}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{width}{plot width}

\item{height}{plot hight}

\item{xlab}{option to change x-axis label}

\item{ylab}{option to change y-axis label}

\item{plot.cex}{cex graphic option}

\item{add}{if TRUE par is not called to enable manual multiplots}
}
\description{
Plots the estimated population trajector relative to GL
}
