% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_plots.R
\name{jrplot_retrobias}
\alias{jrplot_retrobias}
\title{jrplot_retrobias() to plot retrospective pattern}
\usage{
jrplot_retrobias(
  hc,
  output.dir = getwd(),
  as.png = FALSE,
  width = 5,
  height = 4.5,
  add = FALSE,
  xlab = "Year",
  ylab = "Abundance",
  plot.cex = 1,
  Xlim = NULL,
  cols = NULL,
  legend.loc = "topright",
  add.legend = TRUE,
  show.rho = TRUE
)
}
\arguments{
\item{hc}{output from jara_hindast()}

\item{output.dir}{directory to save plots}

\item{as.png}{save as png file of TRUE}

\item{width}{plot width}

\item{height}{plot hight}

\item{add}{if TRUE par is not called to enable manual multiplots}

\item{xlab}{option to change x-axis label}

\item{ylab}{option to change y-axis label}

\item{plot.cex}{plotting parameter}

\item{Xlim}{allows to "zoom-in" requires speficiation Xlim=c(first.yr,last.yr)}

\item{cols}{option to add colour palette}

\item{legend.loc}{location of legend}

\item{add.legend}{option to add legend}

\item{show.rho}{shows rho statistic in plot}
}
\value{
Mohn's rho statistic for several quantaties
}
\description{
Plots retrospective pattern of population trend
}
