% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jara_utils.R
\name{jrroc}
\alias{jrroc}
\title{ROC curve Function}
\usage{
jrroc(x, y, rfp = NULL)
}
\arguments{
\item{x}{observed measure}

\item{y}{true or reference measure}

\item{rfp}{reference point treshold}
}
\description{
ROC curve Function
}
